Complete rewrite
Library has been drastically simplified. It no longer provides tons of entity classes because this should be done in a higher level of an application. This library now directly works with the EDDB JSON types and only provides typescript types for them and some helper functions to stream/read them.